Move Code::Stats HP/MP widget from sidebar into Tech Stack tab
The compact HP/MP bars now live alongside the rest of the Code::Stats
data on the Tech Stack tab instead of cramming the sidebar. Also
shrank the HP/MP pixel-font labels and swapped the explanatory line
to plain text — Press Start 2P reads much larger than its font-size
suggests, so the old caption looked oversized for a two-line sentence
with an inline link.
