+++ src/components/Sidebar.jsx
@@ -1,10 +1,8 @@
 import { useState } from "react"
-import PropTypes from "prop-types"
 import { motion } from "framer-motion"
 import { Cursor, useTypewriter } from "react-simple-typewriter"
 import { FaLinkedin, FaGithub, FaInstagram, FaThreads, FaDiscord, FaFilePdf } from "react-icons/fa6"
 import logo from "../assets/aiman.jpg"
-import useCodeStats from "../hooks/useCodeStats"
 
 const BIRTH_YEAR = 2001
 const BIRTH_MONTH = 12 // December (1-indexed)
@@ -21,31 +19,7 @@ const downloadResume = () => {
     link.click()
 }
 
-const StatBar = ({ id, colorClass, bar }) => (
-    <div className="flex items-center gap-2">
-        <span className={`pixel-font text-[10px] w-5 ${colorClass}`}>{id}</span>
-        <div className="stat-bar-track flex-1">
-            <motion.div
-                className={`stat-bar-fill ${id === "HP" ? "bar-hp" : "bar-mp"}`}
-                initial={{ width: 0 }}
-                animate={{ width: bar ? `${Math.max(bar.pct, 4)}%` : "0%" }}
-                transition={{ duration: 1.2, ease: "easeOut" }}
-            />
-        </div>
-        <span className="text-neutral-500 text-[10px] w-14 text-right tabular-nums">
-            {bar ? `${bar.xp}` : "—"}
-        </span>
-    </div>
-)
-
-StatBar.propTypes = {
-    id: PropTypes.string.isRequired,
-    colorClass: PropTypes.string.isRequired,
-    bar: PropTypes.shape({ xp: PropTypes.number, pct: PropTypes.number }),
-}
-
 const Sidebar = () => {
-    const { hp, mp, loading } = useCodeStats()
     const [copied, setCopied] = useState(false)
 
     const [klass] = useTypewriter({
@@ -103,13 +77,6 @@ const Sidebar = () => {
                         <Cursor cursorColor="#f0c040" />
                         <span className="text-neutral-500"> ]</span>
                     </p>
-                </div>
-
-                {/* HP / MP from Code::Stats */}
-                <div className="rpg-panel-dim p-3 flex flex-col gap-2">
-                    <p className="pixel-font text-[10px] text-gold-400/50 tracking-widest mb-1">LIVE from <a href="https://codestats.net/users/aimxnaim" target="_blank" rel="noreferrer" className="text-gold-400 hover:underline">Code::Stats </a>· HP=TOTAL · MP=TODAY</p>
-                    <StatBar id="HP" colorClass="text-red-400" bar={loading ? null : hp} />
-                    <StatBar id="MP" colorClass="text-blue-400" bar={loading ? null : mp} />
                 </div>
 
                 {/* Socials */}
